requirements.txt
threaded==4.2.0
numpy==1.26.4
pandas==2.2.2
pyserial==3.5
matplotlib==3.8.4
scipy==1.13.0
scikit-learn==1.4.2
